Refactor test structure: individual files per tool per test type
- Split consolidated test files into individual files for better maintainability
- Each tool now has dedicated unit and integration test files
- Removed old consolidated files (source-management.test.ts, source-management-integration.test.ts)
- All 59 tests passing with cleaner organization

Unit tests:
- list-sources.test.ts (4 tests)
- get-source-info.test.ts (3 tests)
- list-source-groups.test.ts (4 tests)
- get-source-group-info.test.ts (4 tests)
- test-connection.test.ts (4 tests)
- query-logs.test.ts (12 tests)

Integration tests:
- list-sources-integration.test.ts (3 tests)
- get-source-info-integration.test.ts (2 tests)
- list-source-groups-integration.test.ts (2 tests)
- get-source-group-info-integration.test.ts (2 tests)
- test-connection-integration.test.ts (2 tests)
- query-logs-integration.test.ts (13 tests)
- source-management-workflows-integration.test.ts (4 tests)
